ecs: Parallel systems (#354)
* ecs: Add ecs_view_chunks api

* ecs: Reduce iterator size

* rend: Force stat collection near the end of the frame

* jobs: Set maximum worker count to 4

Lets wait with spinning up 64 workers until we can actually
meaningly fill them with work :-)

* ecs: Support chunks limit in iterator

* ecs: Refactor ecs_archetype_chunks_non_empty to return u32

* ecs: Support iterators that skip chunks

* ecs: Fix type mismatch in itr

* ecs: Update ecs_view_chunks to only return non-empty chunks

* ecs: Fix bug in chunk skipping

* ecs: Add stepped iterator api

* ecs: Fix typos in view header

* ecs: Combine stepped itr tests

* ecs: Add multi-archetype stepped itr test case

* ecs: Fix bug in stepped iteration

* ecs: Remove redundant assignment from stepped itr

* ecs: Add additional constraints on stepped itrs

* ecs: Only track remaining chunks for stepped itrs

* ecs: Add api to configure the sys parallel count

* ecs: Add order definition tests

* ecs: Use 16 bit for the parallel count

* debug: Remove taskId and workerId from ecs system tab

* jobs: Add JobTaskRange utility struct

* ecs: Refactor runner api with task sets

* ecs: Remove as tracked per system

Won't work anymore if systems can execute on multiple workers

* ecs: Support stats for parallel systems

* ecs: Support dependencies between parallel systems

* ecs: Produce multiple tasks for parallel systems

* ecs: Remove outdated assert from runner

* ecs: Add parallel system test case

* scene: Split skeleton update in 4 parallel tasks

* ecs: Use u16 in stepped iterator api

+++ libs/ecs/src/view_internal.h
@@ -17,7 +17,7 @@ struct sEcsView {
   const EcsViewDef* viewDef;
   EcsStorage*       storage;
   Mem               masks;
-  usize             compCount;
+  u16               compCount;
   DynArray          archetypes; // EcsArchetypeId[] (NOTE: kept sorted)
 };
